Fix mocked InCallService in unit tests.
The InCallServiceFixture essentially mocks what the actual InCallService
does for portions of the Telecom unit tests.  Its implementation did not
actually match how the real InCallService class works.

Updates for calls which don't exist should just be ignored since when the
call is actually added later, the entire call state gets updated.

The likely root cause of this is a change to Telecom to pass along a new
property on calls, which just happens to exercise this code path.

Bug: 118881291
Test: Run telecom tests.
Change-Id: Ie53313ce647a737732bae267a59c619a89036936