Fix BT call state issue
CallAudioManager sets the BT end-of-call tone flag to true after it
calls startTone at the end of a call, no matter whether the call
actually started or not. This causes our call reports to the BT stack to
perpetually be in the disconnecting stack, which apparently is bad.

Fix this so that we only set the flag if the tone actually plays.

Test: unit, manual
Change-Id: I72058667ee6d722fbe5bf87f036a95220cf0a34f
Fixes: 120477414
(cherry picked from commit ee99e09fb328f08d017cd848f0966ee701d3d74b)