Check the ringer playing state before starting
For looping ringtones, we don't currently check whether it's already
playing before starting. If a BT SCO connection starts while the call is
ringing, this code would attempt to start the ringtone twice and cause a
hiccup. This change adds a check to make sure the ringtone isn't already
playing before playing it.

Test: manual
Change-Id: I773182f47e3d815138a252c88271b446b9da02ca
Fixes: 117912905