Fix nuisance report matching for invalid phone numbers.
If a number cannot be converted to E.164 format, it may be null.
Falling back to the unformatted number in this case.  Also change the
where clause to include matching on the raw unformatted number as another
means of matching call log entries.

Test: Added CTS tests
Bug: 123896388
Change-Id: Ia9b2f9ef8159398f0307b3ad04e0a2c532823fda
Merged-In: Ia9b2f9ef8159398f0307b3ad04e0a2c532823fda