Handle video capability changes mid-call.
Add phone account changed listener method in phone account registrar which
will trigger an update of video capabilities for ongoing calls.  This is
to ensure that if the device becomes video capable (or not) during calls
that the caches "is video capable for phone account" indicator for the
calls are updated appropriately and that the current connection capabilities
is updated.

Test: Added new unit tests
Bug: 79185391
Change-Id: If43af6889ff07aad6fd08571a29352d882ffd675