Fix a race condition in BT device handling
When a headset or hearing aid disconnects in favor of the other, we have
a few milliseconds between when BT sends a "device disconnected" for the
first device and when it tells us that the second device is connected.
This causes us to go into a non-BT audio route for a while, but this
sends a "please disconnect audio" message to BT, which disrupts the
setup on the second device.
This change removes the "please disconnect audio" after a BT-initiated
disconnect, since it's redundant anyway.

Test: manual, unit
Bug: 122920967
Change-Id: I5f60b4b68c66c68a73d1fc0dd4e3122b35706d5e