Clear postdial digits on failed outgoing call.
The Postdial digits are populated when a call is created in Telecom; the
connectionservice strips the postdial digits and edits the outgoing
connection back to Telecom once its created, removing them from the
Telecom call.  When a call fails fast in Telephony, a failed connection
is sent to Telecom without the stripped postdial digits; further even if
it were sent the failed connection is simply used as a placeholder to stop
the connection creation process.

To solve this issue, clearing the postdial digits on the telecom call when
a connection creation failure occurs.  The result is the call log entry
will not have postdial digits separated from the dial digits.  However
given this is such an edgecase this workaround is reasonable.

Test: Remove sim and verify this is not an issue.
Bug: 123622080
Change-Id: I78970c747cf83aba8bd9211d7b326ed84433479b