Tweaks to the call blocking options.
1. Add option for device to hide the "block callers not in contacts" option;
there was a large potential for the user to accidentally block too many
calls.  The dialer app would need to do some work to ensure it makes the
user aware of the fact that they're blocking large numbers of calls.
Since this is not mandatory per the CDD requirements, making this a
configurable option.  This way a device can choose to use this feature and
provide a good user experience around exposing the blocked calls.
2. Add an option for a device to combine the block options for "restricted" and
"unknown" callers options.  Conceptually most users won't recognize the
difference between the two types of calls, so it makes sense to combine
them in a generic sense unless there is a strong desire for a device to
support separating these options.

Test: Manual; verified that "not in contacts" option is hidden.
Test: Manual; verified that toggling the option to combine the unknown
and restricted options updates the shared prefs for both.
Fixes: 129353233

Change-Id: I85694512cd71e8296f270d18b0086458dd863e56