Support for audio coupled haptics when playing ringtones.
Updated Ringer.java to suppress system haptics when a ringtone which will
be played has embedded haptics and the device supports audio coupled
haptics.

AsyncRingtonePlayer performs all of its operations on a Handler to avoid
possible strict mode violations when manipulating ringtones, so the play
now returns a completable future which indicates whether audio coupled
haptics will be used to play a ringtone.  Ringer.java waits on this future
and continues with the business of playing the system haptics if needed.

Also updated the RingerTest code because changes made to support ramping
ringtones effectively nullified most of the tests.

Test: Manual, added unit tests
Bug: 127818951
Change-Id: Ie5fc285a0181b27672fd4f8ffae9610b1b1f3b02