Carrier and User-defined call redirection interaction
a) place Carrier call redirection if the intent contains gateway.

b) Telecom should pop out the dialog to users only if: 1) there is
no available carrier redirection service, or 2) there is an available
carrier redirection service, but carrier one does not redirect the call,
and but tells Telecom to place the call unmodified.

Test: Treehugger; manual
Bug: 130193841
Change-Id: I3d44d4666786e7d7b3c7c2bc9bf8f4daa3d3924b