Ensure setHapticsChannelsMuted is set false when haptics should be on.
The behavior of this API has changed; the default behavior is NOT to
play haptics, so we need to explictly turn them on now.
Also fixing the fact that the audio attributes have been previously set
to indicate the content type and usage bits, which would be lost when
setting the haptics flags here.

Test: Verify audio attributes are set appropriately when vibration on.
Test: Verify audio attributes are set appropriately when vibration off.
AOSP: Infeasible due to confidentiality.
Fixes: 130958558
Change-Id: Idc345fdc5a4132e646792f0af1af3efb5b55d0c5