Correct behavior answering incoming call with dialing self-managed call.
When answering an incoming call with a dialing self-managed call in
progress, the incoming call could be answered while the dialing self-mgd
call would remain dialing as well.

Reviewing the logic, dialing calls can't be held, so changing the "canHold"
logic to take that into account.  Now in this scenario the user will be
prompted on the incoming call screen that answering the call will end the
ongoing "X app" call.  If they answer the call, the dialing call is
disconnected.

Test: Reproduced using Telecom test app.  Verified fix corrects issue.
Test: Added unit test for this scenario and ran all existing unit tests.
Bug: 129507552
Change-Id: I80d5ad4875e56df790ad988c01bb7a42a9f0609d
AOSP: aosp/954991