Correct behavior of self-managed incoming call dialog.
The dialog which is shown when there is an incoming self-managed call and
there is one or more ongoing calls in another ConnectionService needed to
have its behavior adjusted.  For incoming self-managed calls, it currently
shows whenever there is another ongoing call in another ConnectionService,
when in reality it ONLY needs to show if the other ongoing calls cannot
be held.

Adjusted the logic to check for this, as well as to check for calls in
another connectionservice based on the package, not based solely on the
targe phone account.  This matches recent changes made to simultaneous
call logic made elsewhere.

Test: Run CTS verifier test.
Bug: 131785694
Change-Id: I22084debef411e55c505931e1be981b2a53c96ca