Make ConnectionManagers multisim aware.
The connection manager functionality was not multisim aware.
To correct this:
1. In PhoneAccountRegistrar#getSimCallManagerFromCall modify so that we
look for the connection manager associated with the subscription of the
call's target phone account.
2. In ConnectionServiceWrapper#queryRemoteConnectionServices modify so
that the calling package is used to find the phone accounts which have
the same subId as the connection manager.  This ensures that the connmgr
can only see the phone accounts for the same sub.

Test: Place call on connection mgr account; verify connection mgr is used.
Test: Place call on regular acct; verify connection mgr is NOT used.
Bug: 131856987
Change-Id: I65be29e1d1700813744a34c213bc380162c5ed2f