Make TelecomManager#addNewIncomingCall DSDS aware.
Modify the isCallerSimCallManager logic in the addNewIncomingCall method to
look up the sim call manager for the target phone account.  This will
ensure that if the target phone account is not the default outgoing call
account that we'll stlil get the correct sim call manager and allow the
caller to add the incoming call.

Bug: 133873542
Test: Manual test with FI incoming WIFI calls; verify it has permission to
add an incoming call when its not the default calling account.

Change-Id: I106f1fe410de976be11e2ee3819092df41894c08