Make addExistingConnection MSIM aware
addExistingConnection was using the default subscription to
find the SIM call manager, which changes, depending on the user
settings for voice subscription ID. In some cases, we were not
finding the Fi call manager (Fi ESIM and no voice pref set), which
was causing addExistingConnection to only try to find a call manager
associated with the pSIM, which does not exist. This was causing
conferencing to fail.

Instead, check all SIM calling accounts for an associated call manager
and make sure the connection being added is associated with one of
those accounts if it is not associated with a SIM calling account.

Test: atest CtsTelecomTestCases; manual testing
Merged-In: Ia69971e089e60a9e908024e883c7e268c650ae97
Change-Id: I72e7d57b651d62ad44356720f03895da0ea9462b