Fix ordering issue with vibration
When someone calls stopRinging, also cancel the pending
completablefuture that's supported to start vibration.

Prior to this change, if a user picked up a call really quickly before
the haptics lookup had a chance to complete, stopRinging would get
called before the vibration even started. Then, when the haptics lookup
completes, it'll trigger the vibration to start. With nothing to stop
it, the vibration will continue until the user reboots the device or
receives another call.

Fixes: 134833743
Test: unit, manual
Change-Id: I6968a1ed0693e0eb839b3d9be032d5bed9525f77