DO NOT MERGE: Running Telecom CallRedirection Dialog on top of InCallActivity
There is a race condition between Telecom CallRedirection Dialog and Dialer
InCallActivity when placing outgoing call. Sometimes, Telecom's dialog goes
first, but sometimes Dialer's UI goes first. If former, Telecom's dialog
cannot be seen on top of InCall UI that prevents the outgoing call to be
placed.

To solve this, start the activity with a small delay.

Test: Manual
Bug: 135140667
Change-Id: Id13c8808e00fc1191d473eb2e1cf73041dd9e31f