Fix call redirection service issues.
1.
Pressing the "home" button or otherwise switching apps while the call
redirection dialog is showing puts the device in a bad state.  The dialog
can no longer be visited from the recents menu, and as such you end up
with a call stuck in a connecting state.  This prevents all future calls
from being placed until reboot.

To fix this, add an onStop handler which will ensure that redirection is
canceled if the user switches apps or hits the home button.

Also, fixed an invalid string which was causing a string formatting error
(too many %s'es).

2. Fix issue where emergency numbers were being routed to a call
redirection service.  These should NOT be.

Test: Place call with redirection app installed; cancel from dialog and
verify it cancels call.
Test: Place call with redirection app installed; place call using carrier
from dialog and verify it connects call via carrier.
Test: Place call with redirection app installed; place call using redir
app from dialog and verify it connects call via redirection app.
Test: Hit home to dismiss the dialog; verify call is canceled.
Test: Add test emergency number, verify can call emergency number
Fixes: 137223601
Fixes: 137315607
Change-Id: I7274bf5c34d5f119e011e6828b97ed2fe270c303