Fix issue where inprogress self-managed call is not seen by car mode ui.
InCallController had an assumption that we should only try to connect
to the car mode UI when we have to first disconnect the non-car mode ui.
For regular calls this assumption is okay.  However, it is possible that
the car mode ui will also handle self-managed calls, which means we need
to start up the car mode ui even though the regular dialer is not running.

Test: Added unit tests to cover this case.
Test: Used Android auto desktop head unit and test telecom app to
generate a selfmanaged call; verified that the fix corrects the issue
and auto sees the call in question when it starts up.
Bug: 137776943
Fixes: 154245428
Merged-In: Iad9c0d8b68557ff0d2d4acdf1ed0eb1ac7ee56df
Change-Id: Ibbf6cf98dfa4fe0a724e1a6ca785184dffaeefb9

Change-Id: Iaa6b1a55ae90e8c6ef94e6930f63f1bd329498d4