Change order of calls to AudioManager
There may sometimes be a delay when calling AudioManager#setMode. If
audio focus is abandoned prior to setting the mode to MODE_NORMAL, there
may be some churn in the audio stack that causes errant noises to be
emitted. Fix this by setting the mode prior to abandoning audio focus.

Fixes: 133197870
Test: manual
Change-Id: I71391abb07ece96bb76056ab8304c68adddbd22a