User cannot reject a waiting call after a switch call fails
Incoming call is set to ANSWERED state if user attempts to answer,
but ANSWERED state is not reset if answer fails due to fail of switch
request.
If switch call fails, reset all ANSWERED calls back to RINGING so that
the user can attempt the operation again.

Test: Manual - Verify that incoming call can be rejected after switch
call is failed.
Bug: 143918738

Change-Id: I8f9cc677ee769f1ea261ede6595850df7de91cb9