Add getDefaultDialerPackage which allows a user id to be specified.
Add new system API which is used from Telephony when we want to figure
out the dialer for a specific user for the purpose of showing the
voicemail notification.
Add new permission check when adding an existing connection with a
specified conference membership.

Bug: 141576016
Test: Manual smoke test.
Test: Run unit tests.
Test: Run CTS tests.
Change-Id: Id0c2a44519bd44705cd8f63553ce2d76442f0841
Merged-In: Id0c2a44519bd44705cd8f63553ce2d76442f0841