Restrict calling in favor of emergency calling - Incoming Cases
The first series of CLs (1/2) that allow for better interactions
between emergency calls and incoming call types. This includes:
1) Disconnecting incoming calls and marking them as MISSED if an
outgoing emergency call is placed before the call is set to active.
2) Not allowing the pulling of external calls while an emergency
call is active and showing a Toast.
3) If a call is in AUDIO_PROCESSING state, hang it up in favor of
a new outgoing emergency call (and mark disconnected call as
MISSED if this is during an initial call screen).
4) If a call is in SIMULATED_RINGING state, hang it up in favor of
a new outgoing emergency call.

Bug: 138741228
Test: manual; atest TelecomUnitTests; atest CtsTelecomTestCases
Change-Id: I74d2b174121fb9c6402caf806d20eb6c5226873a