Support for multiple car-mode InCallServices.
Add support for the new UiModeManager prioritized car mode API in Telecom.
Modified the InCallController so that it can support changes to the current
car mode ICS.
Removed some old companion app code that isn't needed.
Removed car mode override code that isn't needed since we're not using
RoleManager for this.
Added CarModeTracker class to encapsulate tracking which app is in car
mode; this is used by InCallController to manage the current car mode
state.

Test: Created new unit tests and updated existing.
Test: Added new CTS tests.
Test: Manual verification with test apps.
Bug: 144345414
Change-Id: I4cb5c1616cc54dfda97ea1f59271bff57eb1c2ff