Switch ringtone looping to use MediaPlayer looping.
A few releases ago we wanted to switch ringtone looping off of the current
handler-based looping approach.  Because the change came in late, we
have to guard it by a flag and target it only to a single device.  There
is no reason to keep around the legacy handler-based looping and I suspect
this "repeat" could be the source of some bugs where a device fails to
ring.

Test: Manual test; call phone and verify ringtone loops.
Bug: 144345414
Change-Id: I20b74115bfe6776b3ab1c9d763a314bba7f27fab