Fix issue with call redirection activity.
The call redirection activity fails to show on top of the dialer when
the dialer launches just after the confirm dialog activity starts.
Removing the intermediary call redirection activity allows us to use
the SYSTEM_ALERT_WINDOW type for the dialog, ensuring it shows up on top
of the dialer window in all instances.

Test: Manually reproduced bug and verified that the new implementation
does not suffer from race conditions indicated above.
Bug: 137223601
Change-Id: I503ce9f23606ee8fa387fa8924aaeb51173f5571