Don't start a call until after it's been evaluated
Telecom grants the system dialer location permissions if a 3pty dialer attempts
to make an emergency call. However, if a 3pty dialer tries to make an emergency
call, we launch the system dialer's dial activity and terminate the call that
was already started. After termination, we'll revoke the location permission,
thereby killing the system dialer that we just launched.

To fix this, reorder things in CallIntentProcessor so that
evaluateCall() in NewOutgoingCallIntentBroadcaster gets called before we
kick things off in CallsManager, and so that we can quit before we start
binding/setting up stuff related to the call.

Fixes: 152348036
Test: manual, existing unit tests
Change-Id: I0851b730e3d415fbf995a47cf8df60df27add85c