Fix issue where >2 party conference connect time is wrong on child drop.
In the past, ParcelableCallUtils would report the connect time of the
conference to the incall UI by looking for the earliest connect time among
the children of the conference.  This is problematic as it means that if
the oldest participant in a conference disconnects, the UI will now show
the conference connect time of the next latest participant.

To fix this, as children are added to the conference we do the connect
time update on the conference itself and don't do the calculation in the
parcelable call utils.

Test: manual test with carrier calls to verify connect time on the parent
does not get reduced.
Test: Add unit test to verify behavior.
Fixes: 152611824

Merged-In: I94bca6dd6f8379099373bf9ea472a8da1837d9df
Change-Id: Ic5fd2dfef5e0e196fc5d83a216fc7fec469e37f8