Modify BT behavior when connecting to active device
When we try to connect to a device that's already recognized as active
by the BT stack, don't just ignore it -- do some resolution logic to
align everything with what's actually going on in the system.

This change is needed because we may end up in a weird situation with
hearing aids, where BluetoothRouteManager will report AudioOff while the
hearing aid device is still active. This currently happens because we
don't treat active hearing aids as audio-connected outside a call, and
it's not getting updated when a call starts.

This change doesn't fix the root of the problem, but we're planning on
changing all the bluetooth logic in the next release anyway, so this
patch will be sufficient for now.

Test: manual
Test: atest TelecomUnitTests
Fixes: 152724391
Change-Id: Id49dea147d766f45efd5ef2e719f2814b2769ac6