Remove erroneous permission checks in ConnectionServiceWrapper.
These permission checks are not required as the APIs in question are
public APIs so should not requite a system permission at all.

The permission check on addExistingConnection was causing a VERY strange
behavior with a connection manager.  When the connection manager tried to
add a wrapped conference (existing connection), the request was just
ignored.  As the original calls (wrapped) disconnected, the count of calls
in the Connection Manager's connection service dropped to zero, causing
an unbind from the connection manager, causing it to drop all it's calls,
causing the conference to drop.

Fixes: 152094657
Test: Manual test with connection manager; perform VoLTE conference call
and verify that the conference establishes and remains connected as
expected.

Change-Id: If4f210e6523309a8d81af6a82c1f4db0d62f3894