Fix issue where post-dial digits are not re-added to redirected numbers
Post dial digits are stripped prior to redirection, but they were not
being re-added after redirection completed, so dialed calls would not
be able to auto-dial conference bridge numbers (for example).

Also, fix a bug in CallsManager where a null check was missing.

Fixes: 152172598
Test: Manual test using test call redirection all; verify post-dial digits
are dialed to remote end over test.
Test: Added new unit test for CallRedirection processor.
Test: Added new CTS test to verify passing of post-dial digits.

Change-Id: I48c5d895764f0fc0f34c34692d5ab803abe4c5a7