Remote connection and request logging improvements.
In ConnectionServiceWrapper and InCallAdapter, using the "owner info"
field which can be passed via startSession as a way to indicate the
package name of the request originator.  This was already being done for
incallservice requests in full package name format.  For both of these,
standardized on an abbreviated package name which is formed from the first
character of each part of the package name.  E.g. com.android.phone
becomes cap.

Also, when EXTRA_REMOTE_PHONE_ACCOUNT is specified on a call or conference,
ensuring we store that information in the com.android.server.telecom.Call.
This is used when generating the dumpsys logs to ensure it is very clear
when remote connection services are used.

Test: Manual testing of logging
Fixes: 154353494
Change-Id: Ida87c11263232800aa683bb7ab0d18cb4e41c0b1