Properly restart ringtone when BT device connected
A previous change stoppped the ringer from starting after a HFP device
connects when audio focus was previously acquired. This behavior is
partially incorrect, since sometimes we can acquire audio focus without
starting a ringtone. The result of the incorrect behavior is that an
in-band ringtone will fail to play over the HFP device if the device is
on vibrate.

The fix here is to gate the second try on both acquisition of audio
focus and starting the ringtone.

Test: unit tests, manual
Fixes: 160109967
Change-Id: I6c184795b29e09765bfdba300f83cb9471f30751