Fix a car mode tracking leak upon uninstall
When a package gets uninstalled while it's requesting car mode, a race
condition might lead Telecom to receive its car mode cancellation later
than the uninstall, at which point the check to see whether it's a valid
InCallService would fail. If this happens, CarModeTracker will leak the
package and it'll take a restart to clear.

This hasn't been observed in the wild, but it's been causing some
flakiness in Telecom's CTS tests.

Bug: 162647577
Test: atest TelecomUnitTests CtsTelecomTestCases
Change-Id: I549e3d4fe14ad40b89d788d7304821b16764d642