Add stubs in InCallController to track microphone and camera use.
For microphone use we basically track when we bind and unbind to a
calling UX.  This is not perfect as we will probably also include
external calls.

For camera use, we track when the VideoProvider associated with a call
sets or unsets the camera.  This is a pretty reliable indicator of when
the camera is in use as setting the camera from the UX over the
VideoProvider causes the IMS stack to start up the camera, and nulling it
will close the camera.

Change-Id: Ic4fb90f50fa94ccb73e711021ff55bc83a1cb7ed