Add Telecom Call Audio Test App.
It is sometimes useful to be able to have a device auto answer and play
back some audio.  This can be useful for edge of service call testing
where we need one endpoint to remain in a stable service area and another
to be mobile for testing purposes.

This app has one purpose only; it exposes an activity which allows the user
to enable auto answer.  When enabled, the first paragraph of
"A Tale of Two Cities" by Charles Dickens is read to the caller in a
continual loop.

Test: This is just a test app itself.
Bug: 163085177
Change-Id: I46094262f464d935f1d41aa9ac1fe5be65ca43d3