Track disabled car mode ICS in InCallController when UI mode changed.
Currently we only add enabled car mode ICS to the CarModeTracker when we
received the notification of an app entered car mode. However, if the
app disabled before we try to get the package information, we'll ignore
this package forever and can't switch to the car mode app even if it's
been enabled later.

Bug: 186184781
Test: atest CtsTelecomTestCases:CarModeInCallServiceTest
Change-Id: I7d12a97dae2dbfa6373fb201741baafe70bbfd20