Handle TelephonyManager RuntimeExceptions to prevent Telecom crash.
Telephony Manager can throw IllegalStateException when telephony is down
during method invocations.  Although we're catching that, we're NOT
handling the fact that many of the TelephonyManager APIs can ALSO
rethrow a RemoteException as a RuntimeException, causing Telecom to crash.

This was causing intermittent failures in some of the Telecom CTS tests.
Adding catch blocks for ISE which was missing in one place and then the
RuntimeException which was missing in most places.  This is mostly due
to the emergency number APIs.

Test: Build - this is intermittent; this is a speculative fix.
Fixes: 187745038
Change-Id: Idf1640217e9c37f33849fffc4154d070f124c52e