Correct issue where ringtone is not heard on BT headset in vibrate mode.
To correct this issue, listening to STREAM_MUTE_CHANGED_ACTION in
AudioManager and triggering re-evaluation of ringtone playback.
Used existing CallAudioManager#onRingerModeChange() method which does
this for other cases already.
This has the nice side effect of cleaning up the long-standing behavior
where the device will not start ringing if you increase the ring stream
volume while ringing is in progress.

Test: Added new unit test to verify the broadcast will trigger ringtone
re-evaluation.
Test: Manual test; verify that if I unmute the ring stream while a call
is ringing, the ringtone starts playing.
Test: Manual test; verify with device in vibrate-only mode that paired BT
headset has audible ringing.  Confirmed in logs that new code path is hit
which caused ringtone to start.
Fixes: 171545132

Change-Id: I638a6f403fba4a8eff0bd9cf7ade0e8ff6afbc1c