Fixing issue with adding existing connections which start in a conference.
When adding an existing connection to Telecom, Telephony uses an @hide
version of the API which allows the connection's parent to be specified
at the same time.  This was problematic for IMS conference calls wrapped
by a connection manager because they would NEVER get set as part of the
conference.

To fix this, in the RemoteConnectionService we stash the associated parent
for the new existing connection in the connection extras, and we use it
here to restore the desired parent in Telecom when adding the existing
connection.

Also, added some logs to that it is possible to see which calls another
call are eligible to be conferenced with.

Test: Manual testing of IMS conference calls via remote connection service
API
Bug: 188420526
Change-Id: I7b055e2f47ec79235cef08ac3e2e6cbc1ae13e12

Change-Id: Ie0a80fe29cb6bfa62d5d138111fec6b94628d7ee