Propagate verstat from extras if present.
This is used to work around a bug in the remote connection service API
where this property is not being passed through to Telecom.  We
stash the verstat in the connection extras in telephony and make sure to
set it on the Telecom call if a change is detected.

Test: Use connection manager carrier and verify that VERSTAT is now
propagated to Telecom.
Bug: 192608283

Change-Id: I1e0663b5a78ae369af417f32ffc3cdf4a13ba317