Always stop Ringtone instance in AsyncRingtonePlayer
Make sure that any instace of Ringtone created by AsyncRingtonePlayer
will be stopped, which will cause the MediaPlayer to be released, before
replacing it with a new one.

The ringtone created by this class is always a looping one, so if the
ringtone is not stopped it will keep playing on the background.

Bug: 201440544
Fix: 203875109
Test: manual
Change-Id: I716264babe328c817b1a452436a2e1761b27c71d