Filter accounts without emergency ability
If a phone account doesn't expose the ability to
place emergency calls it should be excluded from the
possible list of emergency accounts.

The same check was placed each time an account was being used to
place a call, but who knows what might happen if the code
handling all the available phone accounts happens to crash.

Let's be safe and remove them all before doing any kind of work
instead of missing a 911/112 call.

Signed-off-by: Luca Stefani <luca.stefani.ge1@gmail.com>
Change-Id: I5dfde21b739f96e71c778bdd3f72c9f91af8099c